feat: refactor API and error handling utilities, update notification duration, and improve localization messages
