+++ src/services/auth/auth.service.ts
@@ -10,7 +10,7 @@ import {
   ResetPasswordPayload,
 } from '@/models/schema';
 import { API_RESPONSE, API_SINGLE_RESPONSE } from '@/models/types';
-import { axiosErrorHandler } from '@/utils/axios';
+import { axiosErrorHandler } from '@/utils/axiosUtils';
 
 export const LoginService = async (loginPayload: LoginPayload) => {
   try {
