+++ src/styles/modules/activity.module.scss
@@ -44,7 +44,7 @@
       margin-bottom: 20px;
 
       svg{
-        margin-top: 1px;
+        margin-top: 4px;
         fill: $primary-color;
         font-size: 18px;
       }
